LayoutTests/fast/multicol/newmulticol/balance6.html

<!DOCTYPE html>
<html>
  <head>
    <title>auto-height multicol</title>
    <script src="../../../resources/js-test-pre.js"></script>
    <script>
      if (window.testRunner)
        testRunner.dumpAsText();
      if (window.internals)
        internals.settings.setRegionBasedColumnsEnabled(true);
    </script>
  </head>
  <body>
    <p>There should be no scrollbar on this page.</p>
    <div id="mc" style="-webkit-columns:3; columns:3; overflow:auto; width:500px; orphans:1; widows:1;">
      <div style="height:100px;"></div>
    </div>
    <script>
      shouldBe("mc.offsetHeight", "34");
      shouldBe("mc.scrollWidth", "500");
    </script>
    <script src="../../../resources/js-test-post.js"></script>
  </body>
</html>
